Determine which arraylist is longer and plus them
These two methods are determine which arraylist is longer, and judge for two parts, one is before the dot, and another is after the dot.  If this arraylist is smaller than another one, then add zero util its length is the same with another arraylist.  After that, plus these arraylists numbers in two different arraylists, one is before the dot, and another is after the dot.