Add one to previous grid and print the answer
These methods are determine whether the numbers in the grid of arraylists are over ten, if the number is over, then add one to previous grid, and it also divide two parts, one is before the dot, and another is after the dot.  After that, print the answer of the two numbers.